requirements.txt
beautifulsoup4==4.6.0
Django==2.0.5
djangorestframework==3.8.2
feedparser==5.2.1
newspaper3k==0.2.6
nltk==3.3
psycopg2==2.7.4
python-dateutil==2.7.3
